break out the protocol_storage
