+++ include/glaze/rpc/repe/repe.hpp
@@ -5,6 +5,25 @@
 
 #include "glaze/rpc/repe/header.hpp"
 
+namespace glz
+{
+   // Single template storage for all protocol-specific storage
+   template <uint32_t P>
+   struct protocol_storage
+   {};
+}
+
+namespace glz::detail
+{
+   struct string_hash
+   {
+      using is_transparent = void;
+      [[nodiscard]] size_t operator()(const char* txt) const { return std::hash<std::string_view>{}(txt); }
+      [[nodiscard]] size_t operator()(std::string_view txt) const { return std::hash<std::string_view>{}(txt); }
+      [[nodiscard]] size_t operator()(const std::string& txt) const { return std::hash<std::string>{}(txt); }
+   };
+}
+
 namespace glz::repe
 {
    struct state final
